auth-ui/requirements.txt
Flask>=3.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
